+++ src/lib/npm/outdated.ts
@@ -174,12 +174,15 @@ export const npmOutdated = async (
     const hasLatestUpdate = latest && latest !== info.current
 
     if (hasWantedUpdate || hasLatestUpdate) {
+      const wantedVersion = wanted || info.current
       result.push({
         ...dep,
-        wanted: wanted || info.current,
+        wanted: wantedVersion,
         latest: latest,
         specifier: info.specifier,
         isDevDependency: info.isDevDependency,
+        wantedDate: npmInfo.versionDates?.[wantedVersion],
+        latestDate: npmInfo.versionDates?.[latest],
       })
     }
   })